% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caminhos_pastas.R
\name{caminhos_pastas}
\alias{caminhos_pastas}
\title{Obter Caminho do Projeto por Alias}
\usage{
caminhos_pastas(alias)
}
\arguments{
\item{alias}{Um dos aliases predefinidos: "github", "rpackage", "ciweb",
"cobranca", "extratos", "informakon".}
}
\value{
Uma string com o caminho completo correspondente ao alias fornecido.
}
\description{
Retorna o caminho completo para uma pasta com base no alias fornecido.
}
\examples{
\dontrun{
caminhos_pastas("github")
caminhos_pastas("ciweb")
}

}
