% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_cef_xcef.R
\name{c_cef_xcef}
\alias{c_cef_xcef}
\title{Classify CEF Bank Statement PDF Type}
\usage{
c_cef_xcef(f_caminho.arquivo_c, linhas_c)
}
\arguments{
\item{f_caminho.arquivo_c}{Character string. The file path of the PDF.}

\item{linhas_c}{Character vector. Lines of text extracted from the PDF.}
}
\value{
Character string. The classified type of the CEF bank statement
(e.g., "xcef1", "xcef2", "xcef3", "xcef4", "xcef5", "xcef6", or "desconhecido").
}
\description{
Determines the type of a CEF (Caixa Econômica Federal) bank statement PDF
based on its content.
}
\examples{
# This is an internal function, but an example would look like:
# fake_lines <- c(
#   "Cliente: EMPRESA XYZ",
#   "Conta: 1234 | 001 | 00001234-5",
#   "Data: 01/01/2023 - 10:00",
#   "Mês: Janeiro/2023",
#   "Período: 1 - 31",
#   "Data Mov. Nr. Doc. Histórico Valor Saldo",
#   "01/01/2023 001 PGTO FORNECEDOR 100,00 D 900,00 C",
#   "SAC CAIXA: 0800 726 0101"
# )
# c_cef_xcef("caminho/para/arquivo.pdf", fake_lines)
}
