% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_ana_estq.R
\name{e_ana_estq}
\alias{e_ana_estq}
\title{Extração do Arquivo de Estoque Mais Recente}
\usage{
e_ana_estq(caminho.pasta.inputs_c = caminhos_pastas("fechamento_in"))
}
\arguments{
\item{caminho.pasta.inputs_c}{String do caminho da pasta "Inputs".
Valor padrão: \code{caminhos_pastas("fechamento_in")}.}
}
\value{
Data frame com dados do arquivo de estoque mais recente.
}
\description{
A função e_ana_estq() identifica e extrai o arquivo de estoque mais recente
na pasta especificada.
}
\examples{
\dontrun{
estq_df <- e_ana_estq()
}

}
