% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_cef_cmfcn.R
\name{e_cef_cmfcn}
\alias{e_cef_cmfcn}
\title{Extração dos dados de um arquivo CMF_CN}
\usage{
e_cef_cmfcn(f_caminho.arquivo.cmfcn_c)
}
\arguments{
\item{f_caminho.arquivo.cmfcn_c}{Caminho para o arquivo PDF do CMF_CN.}
}
\value{
Retorna um tibble com os dados extraídos do arquivo CMF_CN.
}
\description{
A função \strong{e_cef_cmfcn} extrai e organiza os dados de um arquivo CMF_CN
em PDF, consolidando informações sobre lançamentos financeiros.
}
\details{
A função lê o arquivo PDF, processa as informações e retorna um tibble
contendo os dados organizados, como contratos, datas, valores e situações.
}
\examples{
\dontrun{
f_caminho.arquivo.cmfcn_c <- "caminho/para/arquivo/MOV_FINANC_CN.pdf"
resultado <- e_cef_cmfcn(f_caminho.arquivo.cmfcn_c)
print(resultado)
}

}
