% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_cef_cmfcns.R
\name{e_cef_cmfcns}
\alias{e_cef_cmfcns}
\title{Consolidação dos dados dos relatórios CMF_CN}
\usage{
e_cef_cmfcns(f_caminho.pasta.ciweb_c)
}
\arguments{
\item{f_caminho.pasta.ciweb_c}{Caminho para a pasta "Relatorios - CIWEB".}
}
\value{
Retorna um tibble consolidado com os dados extraídos dos arquivos CMF_CN.
}
\description{
A função \strong{e_cef_cmfcns} consolida os dados dos relatórios CMF_CN
que estão na pasta "Relatorios - CIWEB".
}
\details{
A função percorre a pasta especificada buscando arquivos PDF que
contenham "MOV_FINANC_CN" no nome. Para cada arquivo encontrado,
chama a função \code{e_cef_cmfcn} para realizar a extração dos dados e,
posteriormente, consolida os resultados em um tibble.
}
\examples{
\dontrun{
f_caminho.pasta.ciweb_c <- "caminho/para/a/pasta/Relatorios - CIWEB"
resultado <- e_cef_cmfcns(f_caminho.pasta.ciweb_c)
print(resultado)
}

}
