% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_cef_cmfcns.R
\name{e_cef_cmfcns_mensal}
\alias{e_cef_cmfcns_mensal}
\title{Agrupamento de CMF_CN por contrato e mês}
\usage{
e_cef_cmfcns_mensal(
  f_caminho.pasta.ciweb_c = caminhos_pastas("ciweb"),
  agrupamento = c("natureza", "lancamentos")
)
}
\arguments{
\item{f_caminho.pasta.ciweb_c}{Caminho para a pasta "Relatorios - CIWEB".}

\item{agrupamento}{Tipo de agrupamento: "natureza" ou "lancamentos".}
}
\value{
Tibble com uma linha por contrato e colunas mensais de valores.
}
\description{
Agrupa os dados de e_cef_cmfcns() por contrato, pivotando valores mensais.
}
