% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_cef_dcd_resumo.R
\name{e_cef_dcd_resumo}
\alias{e_cef_dcd_resumo}
\title{Extração dos dados de um relatório DCD da CEF}
\usage{
e_cef_dcd_resumo(f_caminho.arquivo_c)
}
\arguments{
\item{f_caminho.arquivo_c}{Caminho para o arquivo PDF do DCD.}
}
\value{
Um tibble reunindo as variáveis extraídas (números,
datas de assinatura, codificações internas, etc.).
}
\description{
A função \strong{e_cef_dcd_resumo()} extrai e organiza dados de um relatório
DCD (Demonstração de Cronograma de Desembolso) em PDF da CEF,
consolidando informações e retornando um tibble com os valores
encontrados (custo de obra, saldos, suplementações e detalhes do contrato).
}
\details{
São utilizadas funções da \pkg{pdftools} para leitura do PDF, além de
expressões regulares para filtrar linhas com informações específicas.
}
\examples{
\dontrun{
dcd_res <- e_cef_dcd_resumo("C:/caminho/para/o/arq.pdf")
}

}
