% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_cef_dcds_resumo.R
\name{e_cef_dcds_resumo}
\alias{e_cef_dcds_resumo}
\title{Extração e Consolidação de Relatórios DCDS da CEF}
\usage{
e_cef_dcds_resumo(f_caminho.pasta.ciweb_c = caminhos_pastas("ciweb"))
}
\arguments{
\item{f_caminho.pasta.ciweb_c}{Caminho para a pasta "Relatorios - CIWEB".
O valor padrão é \code{caminhos_pastas("ciweb")}.}
}
\value{
Um tibble com os dados compilados dos relatórios DCDS da CEF.
}
\description{
A função \strong{e_cef_dcds_resumo()} consolida os dados de relatórios DCDS na pasta
"Relatorios - CIWEB", retornando em um tibble as informações extraídas.
}
\details{
Identifica todos os arquivos DCDS com final "demonst_cronograma.pdf" e
aplica \code{\link{e_cef_dcd_resumo}} em cada um deles para extrair os dados.
}
\examples{
\dontrun{
dcds_resumo <- e_cef_dcds_resumo()
head(dcds_resumo)
}

}
