% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_cef_ecn.R
\name{e_cef_ecn}
\alias{e_cef_ecn}
\title{Extração dos dados de um arquivo ECN}
\usage{
e_cef_ecn(f_caminho.arquivo_c, xlsx = FALSE)
}
\arguments{
\item{f_caminho.arquivo_c}{Caminho para o arquivo PDF do ECN.}

\item{xlsx}{Lógico. Se \code{TRUE}, salva o resultado em um arquivo .xlsx.}
}
\value{
Uma lista contendo tibbles de dados segmentados em diversos tópicos.
}
\description{
A função \strong{e_cef_ecn} extrai e organiza dados dos arquivos ECN em PDF,
consolidando informações sobre empreendimentos, empréstimos, dados
consolidados e unidades.
}
\details{
A função lê o arquivo PDF, modela as informações e retorna uma lista
com dados de Empreendimento, Empréstimo, Consolidados e Unidades, baseada
no layout do extrato da CEF.
}
\examples{
\dontrun{
f_caminho.arquivo_c <- "meu_arquivo_ecn.pdf"
resultado <- e_cef_ecn(f_caminho.arquivo_c, xlsx = TRUE)
print(resultado)
}

}
