% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_cef_ecns.R
\name{e_cef_ecns}
\alias{e_cef_ecns}
\title{Consolidação dos dados dos relatórios ECN da CEF}
\usage{
e_cef_ecns(f_caminho.pasta.ciweb_c = caminhos_pastas("ciweb"))
}
\arguments{
\item{f_caminho.pasta.ciweb_c}{Caminho para a pasta "Relatorios - CIWEB".
Por padrão, utiliza o caminho relativo baseado na estrutura do projeto.}
}
\value{
Retorna uma lista com os seguintes elementos:
\itemize{
\item ecn_e: Tibble com os dados dos empreendimentos.
\item ecn_pj: Tibble com os dados dos empréstimos.
\item ecn_c: Tibble com os dados consolidados.
\item ecn_u: Tibble com os dados das unidades.
}
}
\description{
A função \strong{e_cef_ecns} extrai e consolida os dados dos relatórios ECN da CEF
que estão na pasta "Relatorios - CIWEB".
}
\details{
A função percorre a pasta especificada buscando arquivos PDF que
contenham "empreendimento_construcao" no nome. Para cada arquivo encontrado,
chama a função \code{e_cef_ecn} para realizar a extração dos dados e,
posteriormente, consolida os resultados em uma lista com quatro elementos:
ecn_e, Empréstimo, ecn_cs e ecn_u.
}
\examples{
\dontrun{
ecns <- e_cef_ecns(
  f_caminho.pasta.ciweb_c = "caminho/para/a/pasta/Relatorios - CIWEB"
)
print(ecns)
}

}
