% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_cef_eprs.R
\name{e_cef_eprs}
\alias{e_cef_eprs}
\title{Extração e Consolidação de Relatórios EPR da CEF}
\usage{
e_cef_eprs()
}
\value{
Retorna um \code{tibble} contendo as informações extraídas dos relatórios EPR.
}
\description{
A função \strong{e_cef_eprs()} extrai e consolida os dados dos relatórios EPR da CEF
que estão na pasta "Relatorios - CIWEB".
}
\details{
O processo busca todos os arquivos EPR com extensão \code{.pdf}, realiza a extração
dos dados e consolida em um único \code{tibble}.
}
\examples{
\dontrun{
e_cef_eprs()
}

}
