% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_cef_xcefs.r
\name{e_cef_xcefs}
\alias{e_cef_xcefs}
\title{Consolidação dos dados dos extratos da CEF}
\usage{
e_cef_xcefs(f_caminho.pasta.extratos_c = caminhos_pastas("extratos"))
}
\arguments{
\item{f_caminho.pasta.extratos_c}{Caminho completo para a pasta que
contém os arquivos PDF dos extratos.}
}
\value{
Retorna um tibble com as seguintes colunas:
\itemize{
\item Data de lançamento: Date.
\item Data de movimento: Date.
\item Documento: Character.
\item Histórico: Character.
\item Valor: Numeric.
\item Saldo: Numeric.
\item conta.interno: Character.
\item Conta: Character.
\item Agência: Character.
\item produto: Character.
\item CNPJ: Character.
\item Cliente: Character.
\item periodo.inicio: Date.
\item periodo.fim: Date.
\item data.consulta: POSIXct.
}
}
\description{
Consolida e processa dados de múltiplos extratos em PDF da CEF,
combinando-os em um único data frame.
}
\details{
A função percorre a pasta especificada buscando arquivos PDF que
contenham os códigos 2429, 2419 ou 2245, ignorando aqueles que contenham
a palavra "fundo". Para cada arquivo encontrado, chama a função
\code{e_cef_xcef} para realizar a extração dos dados e, posteriormente,
consolida os resultados em um único tibble.
}
\examples{
\dontrun{
extratos <- e_cef_xcefs(
  f_caminho.pasta.extratos_c = "caminho/para/a/pasta/dos/extratos"
)
print(extratos)
}

}
