% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_cef_nplpjs.R
\name{e_cef_nplpjs}
\alias{e_cef_nplpjs}
\title{Consolidação dos dados dos relatórios NPL/PJ da CEF}
\usage{
e_cef_nplpjs(
  f_caminho.pasta.ciweb_c = caminhos_pastas("ciweb"),
  empresa = NULL
)
}
\arguments{
\item{f_caminho.pasta.ciweb_c}{Caminho para a pasta "Relatorios - CIWEB".
Por padrão, utiliza o caminho relativo baseado na estrutura do projeto.}

\item{empresa}{Nome da empresa para filtrar os resultados. Se NULL, retorna
todas as empresas.}
}
\value{
Retorna uma tibble com os dados consolidados dos relatórios NPL/PJ, filtrados
por empresa se o parâmetro empresa for fornecido.
}
\description{
A função \strong{e_cef_nplpjs} extrai e consolida os dados dos relatórios NPL/PJ da CEF
que estão na pasta "Relatorios - CIWEB".
}
\details{
A função percorre a pasta especificada buscando arquivos PDF que
contenham "npl_pj" no nome. Para cada arquivo encontrado,
chama a função \code{e_cef_nplpj} para realizar a extração dos dados e,
posteriormente, consolida os resultados em uma única tabela. Se o parâmetro
empresa for fornecido, filtra os resultados apenas para a empresa especificada.
}
\examples{
\dontrun{
# Retorna dados de todas as empresas
nplpjs <- e_cef_nplpjs(
  f_caminho.pasta.ciweb_c = "caminho/para/a/pasta/Relatorios - CIWEB"
)

# Retorna dados apenas da empresa específica
nplpjs_empresa <- e_cef_nplpjs(
  f_caminho.pasta.ciweb_c = "caminho/para/a/pasta/Relatorios - CIWEB",
  empresa = "AMPLA VILA SONIA"
)
}

}
