\name{e_extratos_mapa}
\alias{e_extratos_mapa}
\title{Cria mapa de extratos únicos em planilha Excel}
\usage{
e_extratos_mapa()
}
\value{
Caminho do arquivo Excel criado
}
\description{
Função que cria uma planilha Excel simples com todas as combinações únicas
da tupla (empresa, banco, conta, descricao) encontradas nos extratos
consolidados CEF e ITAÚ.
}
