% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_ik_cmfs.R
\name{e_ik_cmfs}
\alias{e_ik_cmfs}
\title{Extração do relatório CMF (Conta Movimento Financeiro) mais recente}
\usage{
e_ik_cmfs(caminho.pasta.inputs_c = caminhos_pastas("fechamento_in"))
}
\arguments{
\item{caminho.pasta.inputs_c}{String do caminho da pasta de inputs.
Valor padrão: \code{caminhos_pastas("fechamento_in")}.}
}
\value{
\code{tibble} conforme retornado por \code{\link{e_ik_cmf}}.
}
\description{
A função e_ik_cmfs() localiza o arquivo de Conta Movimento Financeiro (CMF)
mais recente na pasta de inputs do fechamento e o extrai chamando
\code{\link{e_ik_cmf}}.
}
\details{
Os arquivos seguem o padrão \code{cmf-<data>.xlsx}, onde \code{<data>} pode
ser \code{AAAA_MM} (mensal) ou um intervalo \code{AAAA_MM_DD-AAAA_MM_DD}.
O arquivo mais recente é determinado pela última data presente no nome.
}
\examples{
\dontrun{
cmf_t <- e_ik_cmfs()
}

}
\seealso{
\code{\link{e_ik_cmf}}
}
