% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_ik_contrs.R
\name{e_ik_contrs}
\alias{e_ik_contrs}
\title{Extração do Contrato Informakon Mais Recente}
\usage{
e_ik_contrs(caminho.pasta.inputs_c = caminhos_pastas("fechamento_in"))
}
\arguments{
\item{caminho.pasta.inputs_c}{String do caminho da pasta "Inputs".
Valor padrão: \code{caminhos_pastas("fechamento_in")}.}
}
\value{
Data frame com dados dos contratos do arquivo mais recente.
}
\description{
A função e_ik_contrs() identifica e extrai apenas o arquivo de contrato
(contr) mais recente na pasta especificada.
}
\examples{
\dontrun{
# Chamando a função
contratos_df <- e_ik_contrs()
}

}
