% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_ik_cr.R
\name{e_ik_cr}
\alias{e_ik_cr}
\title{Extração de Contas a Receber Informakon}
\usage{
e_ik_cr(f_caminho.pasta.ik_c = caminhos_pastas("fechamento_in"), xlsx = FALSE)
}
\arguments{
\item{f_caminho.pasta.ik_c}{String do caminho da pasta "fechamento_in".
Valor padrão: \code{caminhos_pastas("fechamento_in")}.}

\item{xlsx}{Logical. Se \code{TRUE}, cria um arquivo xlsx com os dados extraídos.
Valor padrão: \code{FALSE}.}
}
\value{
Data frame com dados das contas recebidas consolidadas.
}
\description{
A função e_ik_cr() extrai os dados de contas recebidas dos arquivos na pasta
"fechamento_in", preenche-os em uma planilha xlsx (opcional) e os retorna em
um data frame.
}
\examples{
\dontrun{
# Chamando a função
cr_df <- e_ik_cr()
}

}
