% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_ik_desp.R
\name{e_ik_desp}
\alias{e_ik_desp}
\title{Extração de Despesas Informakon}
\usage{
e_ik_desp(f_caminho.pasta.ik_c = caminhos_pastas("fechamento_in"))
}
\arguments{
\item{f_caminho.pasta.ik_c}{String do caminho da pasta "fechamento_in".
Valor padrão: \code{caminhos_pastas("fechamento_in")}.}
}
\value{
Data frame com dados das despesas consolidadas.
}
\description{
A função e_ik_desp() extrai os dados de despesas dos arquivos na pasta
"fechamento_in" e os retorna em um data frame.
}
\examples{
\dontrun{
# Chamando a função
desp <- e_ik_desp()
}

}
