% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_ik_empr.R
\name{e_ik_empr}
\alias{e_ik_empr}
\title{Extração de Empreendimentos Informakon}
\usage{
e_ik_empr(f_caminho.pasta.ik_c = caminhos_pastas("informakon"))
}
\arguments{
\item{f_caminho.pasta.ik_c}{String do caminho da pasta "informakon".
Valor padrão: \code{caminhos_pastas("informakon")}.}
}
\value{
Data frame com dados dos empreendimentos consolidados.
}
\description{
A função e_ik_empr() extrai os dados de empreendimentos dos arquivos na pasta
"informakon" e os retorna em um data frame.
}
\examples{
\dontrun{
# Chamando a função
empreendimentos_df <- e_ik_empr()
}

}
