% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_ik_fcs.R
\name{e_ik_fcs}
\alias{e_ik_fcs}
\title{Extração do relatório FC (Fechamento de Caixa) mais recente}
\usage{
e_ik_fcs(caminho.pasta.inputs_c = caminhos_pastas("fechamento_in"))
}
\arguments{
\item{caminho.pasta.inputs_c}{String do caminho da pasta de inputs.
Valor padrão: \code{caminhos_pastas("fechamento_in")}.}
}
\value{
\code{tibble} conforme retornado por \code{\link{e_ik_fc}}.
}
\description{
A função e_ik_fcs() localiza o arquivo de Fechamento de Caixa (FC) mais
recente na pasta de inputs do fechamento e o extrai chamando
\code{\link{e_ik_fc}}.
}
\details{
Os arquivos seguem o padrão \code{fc-<data>.xlsx}, onde \code{<data>} pode
ser \code{AAAA_MM} (mensal) ou um intervalo \code{AAAA_MM_DD-AAAA_MM_DD}.
O arquivo mais recente é determinado pela última data presente no nome.
}
\examples{
\dontrun{
fc_t <- e_ik_fcs()
}

}
\seealso{
\code{\link{e_ik_fc}}
}
