% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_ik_rec.R
\name{e_ik_rec}
\alias{e_ik_rec}
\title{Extração de Receitas Informakon}
\usage{
e_ik_rec(f_caminho.pasta.ik_c = c_caminhos_pastas("informakon"), xlsx = FALSE)
}
\arguments{
\item{f_caminho.pasta.ik_c}{String do caminho da pasta "informakon".
Valor padrão: \code{c_caminhos_pastas("informakon")}.}

\item{xlsx}{Logical. Se \code{TRUE}, cria um arquivo xlsx com os dados extraídos.
Valor padrão: \code{FALSE}.}
}
\value{
Data frame com dados das receitas consolidadas.
}
\description{
A função e_ik_rec() extrai os dados de receitas dos arquivos na pasta
"informakon", preenche-os em uma planilha xlsx (opcional) e os retorna em
um data frame.
}
\examples{
\dontrun{
# Chamando a função
receitas_df <- e_ik_rec()
}

}
