% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_ik_unis.R
\name{e_ik_unis}
\alias{e_ik_unis}
\title{Extrair dados de unidades da Informakon}
\usage{
e_ik_unis()
}
\arguments{
\item{caminho.pasta}{String do caminho da pasta "Inputs".
Valor padrão: \code{caminhos_pastas("fechamento_in")}.}
}
\value{
Retorna uma tibble com os dados das unidades organizados.
}
\description{
Extrai e organiza dados de unidades dos arquivos da Informakon.
}
\examples{
\dontrun{
unidades <- e_ik_unis()
print(unidades)
}

}
