% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_ita_xita.R
\name{e_ita_xita}
\alias{e_ita_xita}
\title{Extrai dados de extrato ITAÚ (EXTITA) em formato PDF}
\usage{
e_ita_xita(file_path)
}
\arguments{
\item{file_path}{Caminho para o arquivo PDF do extrato ITAÚ.}
}
\value{
Uma lista com dois data.frames: \code{xita_l} (lançamentos) e \code{xita_c} (saldos).
}
\description{
Função para extrair e estruturar dados de extrato ITAÚ (EXTITA) a partir de um arquivo PDF.
Retorna uma lista com dois data.frames: lançamentos e saldos.
}
