% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_metadados.R
\name{e_metadados}
\alias{e_metadados}
\title{Coleta de Metadados de Diversos Tipos de Arquivos}
\usage{
e_metadados(f_arquivo.tipo_c = NULL)
}
\arguments{
\item{f_arquivo.tipo_c}{O tipo de arquivo para o qual os metadados devem ser coletados.
Valores permitidos: "inad", "viab", "xcef", "xita". Se NULL (padrão),
retorna metadados para todos os tipos.}
}
\value{
Um tibble consolidado com os metadados dos arquivos solicitados.
}
\description{
Função principal que agrega metadados de diferentes fontes.
Pode retornar metadados para um tipo de arquivo específico ou para todos os tipos.
}
