% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_viab.R
\name{e_viab}
\alias{e_viab}
\title{Extracao de dados de um arquivo de viabilidade}
\usage{
e_viab(f_caminho_arquivo_c)
}
\arguments{
\item{f_caminho_arquivo_c}{Caminho local para o arquivo .xlsx de
viabilidade.}
}
\value{
Retorna uma lista nomeada com os seguintes elementos:
\itemize{
\item def: Tibble com indicadores extraidos da aba
"Demonstrativo Eco-Finc".
\item flx: Tibble (placeholder) com dados extraidos da aba "Fluxo".
}
}
\description{
A funcao \strong{e_viab} le um unico arquivo .xlsx de viabilidade e
extrai dados de abas especificas, retornando uma lista nomeada de tibbles.
}
\examples{
\dontrun{
dados_l <- e_viab("caminho/para/Viabilidade Projeto.xlsx")
dados_l$def
dados_l$flx
}

}
