% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_xlsx_dados.r
\name{e_xlsx_dados}
\alias{e_xlsx_dados}
\title{Exporta os dados para um arquivo XLSX}
\usage{
e_xlsx_dados(f_caminho.pasta.dados_c = c_caminhos_pastas("dados"))
}
\arguments{
\item{f_caminho.pasta.dados_c}{Caminho para a pasta onde os dados originais estão
armazenados. Valor padrão: \code{c_caminhos_pastas("dados")}.}
}
\value{
Gera e salva um arquivo XLSX no diretório especificado, com o nome
incluindo a data e a hora da criação.
}
\description{
A função \code{e_xlsx_dados()} transforma os dados obtidos pela função
\code{e_dados()} em um arquivo Excel (.xlsx). Cada aba do arquivo é criada
dinamicamente com os dados de diferentes fontes e configurada com cores e
nomes predefinidos.
}
\details{
A função realiza as seguintes operações:
\itemize{
\item Valida se o diretório com os dados existe.
\item Carrega os dados utilizando \code{e_dados()}.
\item Cria um workbook com abas configuradas conforme a fonte dos dados (por exemplo, "cef" ou "ik")
e atribui cores específicas a cada aba.
\item Salva o arquivo XLSX com um nome composto pela data e hora da geração.
}
}
\examples{
\dontrun{
e_xlsx_dados()
}

}
