% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_edges.R
\name{m_edges}
\alias{m_edges}
\title{Create graph edges for data flow visualization}
\usage{
m_edges()
}
\value{
A list containing the edges_df dataframe
}
\description{
Creates edge definitions for the data flow diagram used with visNetwork
}
