% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_gnw.R
\name{m_gnw}
\alias{m_gnw}
\title{Create data flow graph visualization}
\usage{
m_gnw()
}
\value{
A shiny app representing the data flow
}
\description{
Creates a visNetwork visualization of data flow using nodes_df and edges_df
}
