% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_nodes.R
\name{m_nodes}
\alias{m_nodes}
\title{Create graph nodes and legends for data flow visualization}
\usage{
m_nodes()
}
\value{
A list containing two dataframes:
\itemize{
\item nodes_df: Node definitions for the flow chart
\item nodes.legends_df: Legend definitions for the flow chart
}
}
\description{
Creates nodes and legends dataframes for use with visNetwork
}
