% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_pxr.R
\name{r_pxr}
\alias{r_pxr}
\title{Consolida PXR a partir de viabilidade e fechamento}
\usage{
r_pxr(f_id_pasta_gdrive_c, xlsx = FALSE)
}
\arguments{
\item{f_id_pasta_gdrive_c}{Codigo (ID) da pasta no Google Drive com os
arquivos de viabilidade, repassado para \\code{e_viabs()}.}

\item{xlsx}{Logico. Argumento repassado para \\code{r_fechamento()}.
Padrao: \\code{FALSE}.}
}
\value{
Tibble com a uniao de \\code{e_viabs()$flx} e
\\code{r_fechamento()$mes}.
}
\description{
A funcao \strong{r_pxr} consolida os dados de viabilidade e fechamento
em um unico tibble, realizando apenas um \\code{bind_rows()} entre
\\code{e_viabs()$flx} e \\code{r_fechamento()$mes}.
}
\examples{
\\dontrun{
pxr <- r_pxr("1AbCdEfGhIjKlMnOpQrStUvWxYz")
}

}
