% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_extratos.R
\name{c_extratos}
\alias{c_extratos}
\title{Extrai cobertura de arquivos CEF}
\usage{
c_extratos()
}
\value{
Um tibble com colunas: arquivo, arquivo.subtipo, empresa, conta,
periodo.inicio, periodo.fim, arquivo.tipo
}
\description{
Função para extrair e consolidar informações de cobertura temporal de arquivos
CEF (extratos bancários), retornando um tibble com dados combinados.
}
