% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_cobertura.arquivos.R
\name{e_cobertura.arquivos}
\alias{e_cobertura.arquivos}
\title{Extrai cobertura de arquivos CEF e ITAÚ}
\usage{
e_cobertura.arquivos()
}
\value{
Um tibble com colunas: arquivo, arquivo.subtipo, empresa, conta,
periodo.inicio, periodo.fim, arquivo.tipo
}
\description{
Função para extrair e consolidar informações de cobertura temporal de arquivos
CEF (extratos bancários) e ITAÚ, retornando um tibble com dados combinados.
}
