% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_extratos.R
\name{e_extratos}
\alias{e_extratos}
\title{Extrai e consolida dados de extratos CEF e ITAÚ}
\usage{
e_extratos()
}
\value{
Um tibble com dados consolidados de extratos CEF e ITAÚ, contendo
colunas padronizadas para análise temporal e financeira.
}
\description{
Função para extrair e consolidar dados de extratos bancários CEF e ITAÚ,
combinando-os em um único tibble padronizado.
}
