% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_ik.R
\name{e_ik}
\alias{e_ik}
\title{Extração e Consolidação de Dados Informakon}
\usage{
e_ik(f_caminho.pasta.ik_c = c_caminhos_pastas("informakon"), xlsx = FALSE)
}
\arguments{
\item{f_caminho.pasta.ik_c}{String do caminho da pasta "informakon".
Valor padrão: \code{c_caminhos_pastas("informakon")}.}

\item{xlsx}{Logical. Se \code{TRUE}, cria um arquivo xlsx com os dados extraídos.
Valor padrão: \code{FALSE}.}
}
\value{
Retorna uma lista contendo os dados extraídos (despesas, receitas, etc.) em
diferentes elementos.
}
\description{
A função \strong{e_ik()} extrai os dados dos arquivos na pasta "Informakon",
preenche-os em uma planilha xlsx (opcional) e os retorna em uma lista.
}
\details{
A extração é feita a partir de arquivos na pasta "informakon". Além disso, o
usuário pode optar por criar uma planilha xlsx com os dados processados.
}
\examples{
\dontrun{
# Chamada com valor padrão
dados_informakon <- e_ik()

# Chamada customizada
dados_informakon <- e_ik(
  f_caminho.pasta.ik_c = "C:/caminho/personalizado/informakon",
  xlsx = TRUE
)
}

}
\seealso{
\code{\link{read_excel}}, \code{\link{floor_date}}
}
