% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_ik_cmf.R
\name{e_ik_cmf}
\alias{e_ik_cmf}
\title{Extração de dados CMF Informakon}
\usage{
e_ik_cmf(
  f_caminho.pasta.ik_c = caminhos_pastas("informakon"),
  xlsx = FALSE,
  suprimir_warnings = TRUE
)
}
\arguments{
\item{f_caminho.pasta.ik_c}{String do caminho da pasta "informakon".
Valor padrão: \code{caminhos_pastas("informakon")}.}

\item{xlsx}{Logical. Se \code{TRUE}, cria um arquivo xlsx com os dados extraídos.
Valor padrão: \code{FALSE}.}

\item{suprimir_warnings}{Logical. Se \code{TRUE}, suprime warnings de conversão de tipos.
Valor padrão: \code{TRUE}.}
}
\value{
Data frame com dados do CMF consolidados.
}
\description{
A função e_ik_cmf() extrai os dados de CMF (Conta Movimento Financeiro)
dos arquivos na pasta "informakon", preenche-os em uma planilha xlsx
(opcional) e os retorna em um data frame.
}
\examples{
\dontrun{
# Chamando a função
cmf_df <- e_ik_cmf()

# Com geração de xlsx
cmf_df <- e_ik_cmf(xlsx = TRUE)

# Mostrando warnings de conversão
cmf_df <- e_ik_cmf(suprimir_warnings = FALSE)
}

}
