% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_ik_desp.R
\name{e_ik_desp}
\alias{e_ik_desp}
\title{Extração de Despesas Informakon}
\usage{
e_ik_desp(f_caminho.pasta.ik_c = caminhos_pastas("informakon"), xlsx = FALSE)
}
\arguments{
\item{f_caminho.pasta.ik_c}{String do caminho da pasta "informakon".
Valor padrão: \code{caminhos_pastas("informakon")}.}

\item{xlsx}{Logical. Se \code{TRUE}, cria um arquivo xlsx com os dados extraídos.
Valor padrão: \code{FALSE}.}
}
\value{
Data frame com dados das despesas consolidadas.
}
\description{
A função e_ik_desp() extrai os dados de despesas dos arquivos na pasta
"informakon", preenche-os em uma planilha xlsx (opcional) e os retorna em
um data frame.
}
\examples{
\dontrun{
# Chamando a função
despesas_df <- e_ik_desp()
}

}
