% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_ik_emps.R
\name{e_ik_emps}
\alias{e_ik_emps}
\title{Extração de Empresas Informakon}
\usage{
e_ik_emps(f_caminho.pasta.ik_c = caminhos_pastas("informakon"))
}
\arguments{
\item{f_caminho.pasta.ik_c}{String do caminho da pasta "informakon".
Valor padrão: \code{caminhos_pastas("informakon")}.}
}
\value{
Data frame com dados das empresas consolidadas.
}
\description{
A função e_ik_emps() extrai os dados de empresas dos arquivos na pasta
"informakon" e os retorna em um data frame.
}
\examples{
\dontrun{
# Chamando a função
empresas_df <- e_ik_emps()
}

}
