% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_cobertura.arquivos.R
\name{g_cobertura.arquivos}
\alias{g_cobertura.arquivos}
\title{Plot temporal coverage heatmap for arquivos}
\usage{
g_cobertura.arquivos(cobertura_t = c_extratos())
}
\arguments{
\item{cobertura_t}{Tibble with columns: arquivo, empresa, periodo.inicio, periodo.fim, arquivo.tipo, id}
}
\value{
Plotly heatmap object
}
\description{
Plot temporal coverage heatmap for arquivos
}
