% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_coberturaExtratos.R
\name{g_coberturaExtratos}
\alias{g_coberturaExtratos}
\title{Plot temporal coverage heatmap for extratos}
\usage{
g_coberturaExtratos(cobertura_t = e_cobertura_extratos())
}
\arguments{
\item{cobertura_t}{Tibble with columns: arquivo, empresa, periodo.inicio, periodo.fim, arquivo.tipo, id}
}
\value{
Plotly heatmap object
}
\description{
Plot temporal coverage heatmap for extratos
}
